+++ CHANGELOG.md
@@ -86,6 +86,7 @@ data_source_fiat:
 - Binance parser: added trades statement export format.
 - Conversion tool: added parser for HecoInfo explorer.
 - Conversion tool: added parser for Trezor Suite.
+- Binance parser: added new statement export format.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
 - Binance parser: use filename to determine if deposits or withdrawals.